- Google Maps replaced with Open Street Map to address API key issue - Edge computing of C/N0 and AGC added for real time detection of indicators of jamming/spoofing - GeoPackage writing moved off of Main thread to support more fluid GUI - Initial onboard Sensor integration (just magnetic at this point)
