+++ torgi/build.gradle
@@ -26,6 +26,10 @@ android {
 //            path "CMakeLists.txt"
 //        }
 //    }
+    compileOptions {
+        targetCompatibility 1.8
+        sourceCompatibility 1.8
+    }
 }
 
 dependencies {
@@ -36,7 +40,7 @@ dependencies {
     implementation 'mil.nga.geopackage:geopackage-android:3.0.2'
     implementation 'com.android.support:appcompat-v7:27.1.1'
     implementation 'com.android.support.constraint:constraint-layout:1.1.3'
-    implementation 'com.google.android.gms:play-services-maps:15.0.1'
+    //implementation 'com.google.android.gms:play-services-maps:15.0.1'
     implementation 'com.android.support:design:27.1.1'
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
