+++ torgi/src/main/java/org/sofwerx/torgi/gnss/LatLng.java
@@ -1,0 +1,10 @@
+package org.sofwerx.torgi.gnss;
+
+public class LatLng {
+    public double latitude;
+    public double longitude;
+    public LatLng(double latitude, double longitude) {
+        this.latitude = latitude;
+        this.longitude = longitude;
+    }
+}
