+++ torgi/src/main/java/org/sofwerx/torgi/listener/GnssMeasurementListener.java
@@ -1,6 +1,8 @@
 package org.sofwerx.torgi.listener;
 
 import android.location.GnssMeasurement;
+import android.location.GnssMeasurementsEvent;
+import android.location.GnssStatus;
 import android.location.Location;
 
 import org.sofwerx.torgi.SatStatus;
@@ -12,8 +14,8 @@ import java.util.Collection;
  * This listener is used primarily to report changes detected by TorgiService to any display activity
  */
 public interface GnssMeasurementListener {
-    void onSatStatusUpdated(ArrayList<SatStatus> sats);
+    void onSatStatusUpdated(GnssStatus status);
     void onLocationChanged(Location loc);
-    void onGnssMeasurementReceived(Collection<GnssMeasurement> measurement);
+    void onGnssMeasurementReceived(GnssMeasurementsEvent event);
     void onProviderChanged(String provider, boolean enabled);
 }
