+++ torgi/src/main/java/org/sofwerx/torgi/listener/SensorListener.java
@@ -1,0 +1,7 @@
+package org.sofwerx.torgi.listener;
+
+import android.hardware.SensorEvent;
+
+public interface SensorListener {
+    void onSensorUpdated(SensorEvent event);
+}
