+++ torgi/src/main/res/values/strings.xml
@@ -8,4 +8,8 @@
     <string name="quit_torgi_narrative">Do you want to shutdown TORGI or do you want TORGI to keep collecting data in the background?</string>
     <string name="quit_yes">Shutdown</string>
     <string name="quit_run_in_background">Run in background</string>
+    <string name="does_not_support_battery_optimization">This device does not support battery optimization.</string>
+    <string name="enable_battery_optimization">Settings adjustment recommended</string>
+    <string name="battery_optimizations_narrative">Can I turn off battery optimization for TORGI? This will allow better updates from sensors but will drain the battery a bit more.</string>
+    <string name="battery_optimize_yes">Ok</string>
 </resources>
